// components/Panel/config/tailwindUtilities.generated.ts
export const TAILWIND_UTILITY_SETS = {
// ...
  "fontWeight": {
    "base": {
      control: "select",
      options: [
        { label: "font-thin", value: "font-thin", cssValue: "100" },
        { label: "font-extralight", value: "font-extralight", cssValue: "200" },
        { label: "font-light", value: "font-light", cssValue: "300" },
        { label: "font-normal", value: "font-normal", cssValue: "400" },
        { label: "font-medium", value: "font-medium", cssValue: "500" },
        { label: "font-semibold", value: "font-semibold", cssValue: "600" },
        { label: "font-bold", value: "font-bold", cssValue: "700" },
        { label: "font-extrabold", value: "font-extrabold", cssValue: "800" },
        { label: "font-black", value: "font-black", cssValue: "900" },
      ],
    },
    "md": {
      control: "select",
      options: [
        { label: "md:font-thin", value: "md:font-thin", cssValue: "100" },
        { label: "md:font-extralight", value: "md:font-extralight", cssValue: "200" },
        { label: "md:font-light", value: "md:font-light", cssValue: "300" },
        { label: "md:font-normal", value: "md:font-normal", cssValue: "400" },
        { label: "md:font-medium", value: "md:font-medium", cssValue: "500" },
        { label: "md:font-semibold", value: "md:font-semibold", cssValue: "600" },
        { label: "md:font-bold", value: "md:font-bold", cssValue: "700" },
        { label: "md:font-extrabold", value: "md:font-extrabold", cssValue: "800" },
        { label: "md:font-black", value: "md:font-black", cssValue: "900" },
      ],
    },
    "lg": {
      control: "select",
      options: [
        { label: "lg:font-thin", value: "lg:font-thin", cssValue: "100" },
        { label: "lg:font-extralight", value: "lg:font-extralight", cssValue: "200" },
        { label: "lg:font-light", value: "lg:font-light", cssValue: "300" },
        { label: "lg:font-normal", value: "lg:font-normal", cssValue: "400" },
        { label: "lg:font-medium", value: "lg:font-medium", cssValue: "500" },
        { label: "lg:font-semibold", value: "lg:font-semibold", cssValue: "600" },
        { label: "lg:font-bold", value: "lg:font-bold", cssValue: "700" },
        { label: "lg:font-extrabold", value: "lg:font-extrabold", cssValue: "800" },
        { label: "lg:font-black", value: "lg:font-black", cssValue: "900" },
      ],
    },
  },
// ...
} as const;